+++ console_v2/res/entities/EntityStone.cs
@@ -34,7 +34,7 @@ namespace console_v2.res.entities
         {
             if (!Exists) return;
 
-            var scythes = triggerer.Inventory.Tools.Where(tool => tool.DBRef == NeededTool);
+            var scythes = triggerer.Inventory.Tools.Where(tool => tool.DBRef == (int)NeededTool);
             if (scythes.Count() == 0)
                 return;
             var str = scythes.Max(scythe => scythe.STR);
